Clean singlep.py, added classification loss functions for imbalanced data
