+++ mlreco/models/singlep.py
@@ -1,23 +1,19 @@
-import sys
 import numpy as np
 import torch
 import torch.nn as nn
-import torch.nn.functional as F
 
 from torch_geometric.data import Batch, Data
 
 from mlreco.models.layers.common.cnn_encoder import SparseResidualEncoder
 from mlreco.models.experimental.layers.pointnet import PointNetEncoder
 
-from collections import defaultdict, Counter, OrderedDict
 from mlreco.models.layers.common.activation_normalization_factories import activations_construct
 from mlreco.models.layers.common.configuration import setup_cnn_configuration
 from mlreco.models.experimental.bayes.encoder import MCDropoutEncoder
-from mlreco.models.experimental.bayes.evidential import EVDLoss
-from mlreco.models.layers.cluster_cnn.losses.lovasz import StableBCELoss
-
 from mlreco.utils.gnn.data import split_clusts
 from mlreco.utils.gnn.cluster import form_clusters, get_cluster_label
+
+from mlreco.models.experimental.losses.classification import *
 from mlreco.utils.globals import PID_COL
 
 class ParticleImageClassifier(nn.Module):
@@ -152,232 +148,42 @@ class MultiParticleImageClassifier(ParticleImageClassifier):
         return res
 
 
-class DUQParticleClassifier(ParticleImageClassifier):
+def classification_loss_dict(name):
+    """Loss function dictionary for classification tasks.
+
+    All loss functions assume that the inputs are of the form
+    (logits, labels), where logits is a tensor of shape (N, C) and
+    labels is a tensor of shape (N,) containing the class labels.
+    
+    Also assumes that reduction is set to 'mean' in the loss function.
     """
-    Uncertainty Estimation Using a Single Deep Deterministic Neural Network
-    https://arxiv.org/pdf/2003.02037.pdf
-    Joost van Amersfoort, Lewis Smith, Yee Whye Teh, Yarin Gal.
-
-    Pytorch Implementation for SparseConvNets with MinkowskiEngine backend.
-    """
-
-    MODULES = ['network_base', 'particle_image_classifier', 'mink_encoder']
-
-    def __init__(self, cfg, name='duq_particle_classifier'):
-        super(DUQParticleClassifier, self).__init__(cfg, name=name)
-        self.model_config = cfg.get(name, {})
-        self.final_layer = None
-        self.gamma = self.model_config.get('gamma', 0.999)
-        self.sigma = self.model_config.get('sigma', 0.3)
-
-        self.embedding_dim = self.model_config.get('embedding_dim', 64)
-        self.latent_size = self.model_config.get('latent_size', 256)
-
-        self.w = nn.Parameter(torch.zeros(self.embedding_dim, self.num_classes, self.latent_size))
-
-        nn.init.kaiming_normal_(self.w, nonlinearity='relu')
-
-        self.register_buffer('N', torch.ones(self.num_classes) * 20)
-        self.register_buffer('m', torch.normal(torch.zeros(self.embedding_dim, self.num_classes), 0.05))
-
-        self.m = self.m * self.N.unsqueeze(0)
-
-    def embed(self, x):
-
-        feats = self.encoder(x)
-        out = torch.einsum('ij,mnj->imn', feats, self.w)
-        return out
-
-    def bilinear(self, z):
-        embeddings = self.m / self.N.unsqueeze(0)
-
-        diff = z - embeddings.unsqueeze(0)
-        y_pred = (- diff**2).mean(1).div(2 * self.sigma**2).exp()
-
-        return y_pred
-
-    def forward(self, input):
-
-        point_cloud, = input
-        if self.training:
-            point_cloud.requires_grad_(True)
-
-        z = self.embed(point_cloud)
-        y_pred = self.bilinear(z)
-
-        res = {
-            'score': [y_pred],
-            'embedding': [z],
-            'input': [point_cloud],
-            'centroids' : [self.m.detach().cpu().numpy() / self.N.detach().cpu().numpy()]
-        }
-
-        self.z = z
-        self.y_pred = y_pred
-
-        return res
-
-    def update_buffers(self):
-        with torch.no_grad():
-            # normalizing value per class, assumes y is one_hot encoded
-            self.N = torch.max(self.gamma * self.N + (1 - self.gamma) * self.y_pred.sum(0), torch.ones_like(self.N))
-            # compute sum of embeddings on class by class basis
-            features_sum = torch.einsum('ijk,ik->jk', self.z, self.y_pred)
-            self.m = self.gamma * self.m + (1 - self.gamma) * features_sum
-
-
-class EvidentialParticleClassifier(ParticleImageClassifier):
-
-    MODULES = ['network_base', 'particle_image_classifier', 'mink_encoder']
-    def __init__(self, cfg, name='evidential_image_classifier'):
-        super(EvidentialParticleClassifier, self).__init__(cfg, name=name)
-        self.final_layer_name = cfg.get(name, {}).get('final_layer_name', 'relu')
-        if self.final_layer_name == 'relu':
-            self.final_layer = nn.Sequential(
-                nn.Linear(self.encoder.latent_size, self.num_classes),
-                nn.ReLU())
-        elif self.final_layer_name == 'softplus':
-            self.final_layer = nn.Sequential(
-                nn.Linear(self.encoder.latent_size, self.num_classes),
-                nn.Softplus())
-        else:
-            raise Exception("Unknown output activation name %s provided" % self.final_layer_name)
-        self.eps = cfg.get(name, {}).get('eps', 0.0)
-
-    def forward(self, input):
-        point_cloud, = input
-        out = self.encoder(point_cloud)
-        evidence = self.final_layer(out)
-        # print("Evidence = ", evidence)
-        concentration = evidence + 1.0
-        S = torch.sum(concentration, dim=1, keepdim=True)
-        uncertainty = self.num_classes / (S + self.eps)
-        res = {
-            'evidence': [evidence],
-            'uncertainty': [uncertainty],
-            'concentration': [concentration],
-            'expected_probability': [concentration / S]
-        }
-        return res
-
-
-class BayesianParticleClassifier(torch.nn.Module):
-
-    MODULES = ['network_base', 'mcdropout_encoder']
-
-    def __init__(self, cfg, name='bayesian_particle_classifier'):
-        super(BayesianParticleClassifier, self).__init__()
-        setup_cnn_configuration(self, cfg, 'network_base')
-
-        self.model_config = cfg.get(name, {})
-        self.num_classes = self.model_config.get('num_classes', 5)
-        self.encoder = MCDropoutEncoder(cfg)
-
-        self.mode = self.model_config.get('mode', 'mc_dropout')
-
-        if self.mode == 'evidential':
-            self.logit_layer = nn.Sequential(
-                nn.Linear(self.encoder.latent_size, self.num_classes),
-                nn.Softplus())
-        else:
-            self.logit_layer = nn.Sequential(
-                nn.ReLU(),
-                nn.Linear(self.encoder.latent_size, self.num_classes))
-
-        self.num_samples = self.model_config.get('num_samples', 20)
-        self.eps = self.model_config.get('eps', 0.0)
-        print('Dropout network will run inference on {} mode'.format(self.mode))
-
-
-    def evidential_forward(self, input):
-        point_cloud, = input
-        out = self.encoder(point_cloud)
-        out = self.logit_layer(out) + self.eps
-
-        concentration = out + 1.0
-        S = torch.sum(concentration, dim=1, keepdim=True)
-        uncertainty = self.num_classes / (S + 0.000001)
-
-        res = {}
-
-        res['evidence'] = [out]
-        res['uncertainty'] = [uncertainty]
-        res['concentration'] = [concentration]
-        res['expected_probability'] = [concentration / S]
-
-        return res
-
-
-    def mc_forward(self, input, num_samples=None):
-
-        with torch.no_grad():
-            if num_samples is None:
-                num_samples = self.num_samples
-
-            print("Number of Samples = {}".format(num_samples))
-
-            point_cloud, = input
-
-            device = point_cloud.device
-
-            for m in self.modules():
-                if m.__class__.__name__ == 'Dropout':
-                    m.train()
-
-            num_batch = torch.unique(point_cloud[:, 0].int()).shape[0]
-
-            pvec = torch.zeros((num_batch, self.num_classes)).to(device)
-            logits = torch.zeros((num_batch, self.num_classes)).to(device)
-            discrete = torch.zeros((num_batch, self.num_classes)).to(device)
-
-            eye = torch.eye(self.num_classes).int().to(device)
-
-            for i in range(num_samples):
-                x = self.encoder(point_cloud)
-                out = self.logit_layer(x)
-                logits += out
-                pred = torch.argmax(out, dim=1)
-                pvec += F.softmax(out, dim=1)
-                discrete += eye[pred]
-
-            mc_dist = discrete / float(num_samples)
-            softmax_probs = pvec / float(num_samples)
-            logits = logits / float(num_samples)
-            # logits = torch.logit(softmax_probs)
-            res = {
-                'softmax': [softmax_probs],
-                'logits': [logits],
-                'mc_dist': [mc_dist]
-            }
-            return res
-
-    def standard_forward(self, input, verbose=False):
-        if verbose:
-            sys.stdout.write("Forwarding using weight averaging (standard dropout) ...")
-            sys.stdout.flush()
-        point_cloud, = input
-        out = self.encoder(point_cloud)
-        out = self.logit_layer(out)
-        res = {
-            'logits': [out]
-        }
-        return res
-
-
-    def forward(self, input):
-        if (not self.training) and (self.mode == 'mc_dropout'):
-            return self.mc_forward(input)
-        elif self.mode == 'evidential':
-            return self.evidential_forward(input)
-        else:
-            return self.standard_forward(input)
+    
+    loss_dict = {
+        'cross_entropy': CrossEntropyLoss,
+        'focal': FocalLoss,
+        'dice': DiceLoss,
+        'lovasz_softmax': LovaszSoftmaxLoss,
+        'jaccard': JaccardLoss,
+        'tversky': TverskyLoss,
+        'focal_tversky': FocalTverskyLoss,
+        'log_cosh_dice': LogCoshDiceLoss,
+        'log_dice': LogDiceLoss
+    }
+    
+    constructor = loss_dict[name]
+    
+    print(f"Initialized {constructor.__name__} Loss for Classification.")
+    
+    return constructor
 
 
 class ParticleTypeLoss(nn.Module):
 
     def __init__(self, cfg, name='particle_type_loss'):
         super(ParticleTypeLoss, self).__init__()
-        self.xentropy = nn.CrossEntropyLoss(ignore_index=-1)
+        self.loss_fn_name = cfg.get(name, {}).get('loss', 'cross_entropy')
+        loss_kwargs = cfg.get(name, {}).get('loss_kwargs', {})
+        self.loss_fn = classification_loss_dict(self.loss_fn_name)(**loss_kwargs)
         self.num_classes = cfg.get(name, {}).get('num_classes', 5)
 
     def forward(self, out, type_labels):
@@ -385,7 +191,7 @@ class ParticleTypeLoss(nn.Module):
         logits = out['logits'][0]
         labels = type_labels[0][:, -1].to(dtype=torch.long)
 
-        loss = self.xentropy(logits, labels)
+        loss = self.loss_fn(logits, labels)
 
         pred   = torch.argmax(logits, dim=1)
         accuracy = float(torch.sum(pred[labels > -1] == labels[labels > -1])) / float(labels[labels > -1].shape[0])
@@ -414,9 +220,17 @@ class MultiParticleTypeLoss(nn.Module):
         self.batch_col   = loss_cfg.get('batch_col',   0)
         self.target_col  = loss_cfg.get('target_col', 9)
         self.balance_classes = loss_cfg.get('balance_classes', False)
+        self.loss_fn_name = loss_cfg.get('loss', 'cross_entropy')
+        
+        loss_kwargs = loss_cfg.get('loss_kwargs', {})
+        
+        if self.loss_fn_name != 'cross_entropy' and self.balance_classes:
+            raise ValueError('Only cross entropy loss can be balanced.')
 
-        reduction = 'mean' if not self.balance_classes else 'sum'
-        self.xentropy = nn.CrossEntropyLoss(ignore_index=-1, reduction=reduction)
+        if self.balance_classes:
+            self.loss_fn = nn.CrossEntropyLoss(ignore_index=-1, reduction='sum')
+        else:
+            self.loss_fn = classification_loss_dict(self.loss_fn_name)(**loss_kwargs)
 
         self.split_input_mode = loss_cfg.get('split_input_as_tg_batch', False)
 
@@ -446,7 +260,9 @@ class MultiParticleTypeLoss(nn.Module):
         if not len(labels):
             res = {
                 'loss': torch.tensor(0., requires_grad=True, device=valid_labels.device),
-                'accuracy': 1.
+                'accuracy': 1.,
+                'reg_loss': 0.0,
+                'ce_loss': 0.0
             }
             for c in range(self.num_classes):
                 res[f'accuracy_class_{c}'] = 1.
@@ -454,144 +270,37 @@ class MultiParticleTypeLoss(nn.Module):
 
         labels = torch.tensor(np.concatenate(labels), dtype=torch.long, device=valid_labels.device)
         logits = torch.cat(logits, axis=0)
+        
+        res = {
+            'loss': torch.tensor(0., requires_grad=True, device=valid_labels.device),
+            'accuracy': 1.,
+            'reg_loss': 0.0,
+            'ce_loss': 0.0
+        }
 
         if not self.balance_classes:
-            loss = self.xentropy(logits, labels)
+            output = self.loss_fn(logits, labels)
+            loss = output['loss']
+            res['reg_loss'] = output['reg_loss']
+            res['ce_loss'] = output['ce_loss']
         else:
             classes, counts = labels[labels>-1].unique(return_counts = True)
             weights = torch.sum(counts)/counts/self.num_classes
             loss = 0.
             for i, c in enumerate(classes):
                 class_mask = labels == c
-                loss += weights[i] * self.xentropy(logits[class_mask], labels[class_mask]) / torch.sum(counts)
+                loss += weights[i] * self.loss_fn(logits[class_mask], labels[class_mask]) / torch.sum(counts)
 
         pred   = torch.argmax(logits, dim=1)
         accuracy = float(torch.sum(pred[labels > -1] == labels[labels > -1])) / float(labels[labels > -1].shape[0])
-
-        res = {
-            'loss': loss,
-            'accuracy': accuracy
-        }
+        
+        res['loss'] = loss
+        res['accuracy'] = accuracy
 
         for c in range(self.num_classes):
             mask = labels == c
             res[f'accuracy_class_{c}'] = \
                 float(torch.sum(pred[mask] == labels[mask])) / float(torch.sum(mask)) \
                 if torch.sum(mask) else 1.
-
-        return res
-
-
-class MultiLabelCrossEntropy(nn.Module):
-
-    def __init__(self, cfg, name='duq_particle_classifier'):
-        super(MultiLabelCrossEntropy, self).__init__()
-        self.xentropy = nn.BCELoss(reduction='none')
-        self.num_classes = 5
-        model_cfg = cfg.get(name, {})
-        self.grad_w = model_cfg.get('grad_w', 0.0)
-        self.grad_penalty = model_cfg.get('grad_penalty', True)
-
-
-    @staticmethod
-    def calc_gradient_penalty(x, y_pred):
-        '''
-        Code From the DUQ main Github Repository:
-        https://github.com/y0ast/deterministic-uncertainty-quantification
-
-        Author: Joost van Amersfoort
-        '''
-        gradients = torch.autograd.grad(
-                outputs=y_pred,
-                inputs=x,
-                grad_outputs=torch.ones_like(y_pred),
-                create_graph=True,
-            )[0]
-
-        gradients = gradients.flatten(start_dim=1)
-
-        # L2 norm
-        grad_norm = gradients.norm(2, dim=1)
-
-        # Two sided penalty
-        gradient_penalty = ((grad_norm - 1) ** 2).mean()
-
-        # One sided penalty - down
-    #     gradient_penalty = F.relu(grad_norm - 1).mean()
-
-        return gradient_penalty
-
-    def forward(self, out, type_labels):
-        # print(type_labels)
-        probas = out['score'][0]
-        device = probas.device
-        labels_one_hot = torch.eye(self.num_classes)[type_labels[0][:, 0].long()].to(device=device)
-        loss1 = self.xentropy(probas, labels_one_hot)
-        pred = torch.argmax(probas, dim=1)
-        labels = type_labels[0][:, 0].long()
-
-        # Comptue gradient penalty
-        loss2 = 0
-        if self.grad_penalty:
-            loss2 = self.calc_gradient_penalty(out['input'][0], probas)
-
-        loss1 = loss1.sum(dim=1).mean()
-        loss = loss1 + self.grad_w * loss2
-
-        accuracy = float(torch.sum(pred == labels)) / float(labels.shape[0])
-
-        res = {
-            'loss': loss,
-            'loss_embedding': float(loss1),
-            'loss_grad_penalty': float(loss2),
-            'accuracy': accuracy
-        }
-
-        print(res)
-
-        acc_types = {}
-        for c in labels.unique():
-            mask = labels == c
-            acc_types['accuracy_class_{}'.format(int(c))] = \
-                float(torch.sum(pred[mask] == labels[mask])) / float(torch.sum(mask))
-        return res
-
-
-class EvidentialLearningLoss(nn.Module):
-
-    def __init__(self, cfg, name='evidential_learning_loss'):
-        super(EvidentialLearningLoss, self).__init__()
-        self.loss_config = cfg.get(name, {})
-        self.evd_loss_name = self.loss_config.get('evd_loss_name', 'edl_sumsq')
-        self.num_classes = self.loss_config.get('num_classes', 5)
-        self.num_total_iter = self.loss_config.get('num_total_iter', 50000)
-        self.loss_fn = EVDLoss(self.evd_loss_name, 'mean', T=self.num_total_iter)
-
-    def forward(self, out, type_labels, iteration=0):
-
-        alpha = out['concentration'][0]
-        probs = out['expected_probability'][0]
-        device = alpha.device
-
-        labels = type_labels[0][:, 0].to(dtype=torch.long)
-
-        labels_onehot = torch.eye(self.num_classes, device=device)[labels]
-
-        loss_batch = self.loss_fn(alpha, labels_onehot, t=iteration)
-        loss = loss_batch.mean()
-        pred = torch.argmax(probs, dim=1)
-
-        accuracy = float(torch.sum(pred == labels)) / float(labels.shape[0])
-
-        res = {
-            'loss': loss,
-            'accuracy': accuracy
-        }
-
-        acc_types = {}
-        for c in labels.unique():
-            mask = labels == c
-            acc_types['accuracy_class_{}'.format(int(c))] = \
-                float(torch.sum(pred[mask] == labels[mask])) / float(torch.sum(mask))
 
         return res
